fix fmt issue see https://github.com/fmtlib/fmt/issues/391#issuecomment-1039276864

+++ velox/common/compression/Compression.cpp
@@ -56,7 +56,7 @@ CompressionKind codecTypeToCompressionKind(folly::io::CodecType type) {
     case folly::io::CodecType::GZIP:
       return CompressionKind_GZIP;
     default:
-      VELOX_UNSUPPORTED("Not support folly codec type {}", type);
+      VELOX_UNSUPPORTED("Not support folly codec type {}", static_cast<int32_t>(type));
   }
 }
